src/cofounder_agent/requirements.txt
# AI Co-Founder System Requirements
# Complete dependency list for the AI Co-Founder system including all components

# ===== CORE AI & MCP FRAMEWORK =====
# Model Context Protocol and AI model integration
mcp>=1.0.0
openai>=1.30.0
google-generativeai>=0.8.5
anthropic>=0.18.0

# ===== BUSINESS INTELLIGENCE & ANALYTICS =====
# Required for business intelligence system and data processing
aiohttp>=3.9.0
numpy>=1.24.0
pandas>=2.0.0
python-dateutil>=2.8.0

# ===== WEB FRAMEWORK & API =====
# FastAPI and web server capabilities
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
starlette>=0.27.0

# ===== SECURITY & AUTHENTICATION =====
# Encryption and TOTP authentication
cryptography>=42.0.0
pyotp>=2.9.0

# ===== MEMORY & SEMANTIC SEARCH =====
# Advanced memory system with semantic similarity
sentence-transformers>=2.2.0

# ===== DATABASE & STORAGE =====
# PostgreSQL with asyncpg driver (REQUIRED - no SQLite fallback)
# asyncpg: High-performance async PostgreSQL driver (direct SQL, no ORM)
asyncpg>=0.29.0

# ===== DATA PROCESSING =====
# Data handling, parsing, and web scraping
beautifulsoup4>=4.12.0
requests>=2.32.4
markdown>=3.5.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pypdf>=6.0.0

# ===== TESTING FRAMEWORK =====
# Comprehensive testing infrastructure
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-timeout>=2.1.0

# ===== DEVELOPMENT & MONITORING =====
# Development tools and logging
structlog>=23.2.0
rich>=13.7.0
typer>=0.9.0

# ===== SECURITY & AUTHENTICATION =====
# Security and authentication
cryptography>=41.0.0
PyJWT>=2.8.0
pyotp>=2.9.0
python-jose>=3.3.0
email-validator>=2.1.0
passlib>=1.7.4

# ===== UTILITIES =====
# General utilities and helpers
click>=8.1.0
tqdm>=4.66.0
pyyaml>=6.0.0
urllib3>=2.5.0